requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
langfuse>=2.0.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
PyYAML>=6.0.0
feedparser>=6.0.0
claude-agent-sdk>=0.1.0
pypdf>=4.0.0
